DEV-109: Split cron jobs into one task each
